changed functionality to statemachine
the llearngpio daemon will from now on function as a whole gpio semantic
handling daemon. it will permanently run and analyse the gpio data,
decide in which status the statemachine is and post it to mqtt
afterwards. it will also generate a logfile in a tmpfs. when the machine
has stopped, this logfile will be copied and saved to a file on the
sdCard and be pushed on a backup server.